requirements.txt
huggingface-hub==0.11.1
joblib==1.2.0
nltk==3.8.1
numpy==1.24.0
rdflib==6.2.0
scikit-learn==1.2.0
sentence-transformers==2.2.2
spacy==3.4.4
tensorflow==2.11.0
transformers==4.25.1
click==8.1.3
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==0.21.0
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.11.0
aniso8601==9.0.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
python-dotenv==0.21.0
pytz==2022.7.1
six==1.16.0
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.11.0
aniso8601==9.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7.1
six==1.16.0
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.11.0
aniso8601==9.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
pycparser==2.21
python-dotenv==0.21.0
pytz==2022.7.1
six==1.16.0
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.11.0
